ai-research-assistant/apps/backend-api/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6

sqlalchemy==2.0.36
pydantic==2.10.6
pydantic-settings==2.7.1
email-validator==2.2.0

python-multipart==0.0.20

python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

pytest==8.3.4
httpx==0.27.2